removal of Long types (unnecessary can just use Int); compilation errors fixed; basic cell operations added

+++ demo/src/main/java/com/example/demo/repository/ActivityLogRepository.java
@@ -3,5 +3,5 @@ package com.example.demo.repository;
 import com.example.demo.model.ActivityLog;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ActivityLogRepository extends JpaRepository<ActivityLog, Long> {
+public interface ActivityLogRepository extends JpaRepository<ActivityLog, Integer> {
 }
